Create Are they the "same"?.py
Given two arrays a and b write a function comp(a, b)  that checks whether the two arrays have the "same" elements, with the same multiplicities (the multiplicity of a member is the number of times it appears). "Same" means, here, that the elements in b are the elements in a squared, regardless of the order.